Context version for OpenCL

+++ src/core/Kernel-impl.h
@@ -22,7 +22,7 @@ KernelConfig::~KernelConfig()
 }
 
 
-inline unsigned
+inline size_t
 KernelConfig::argsSize() const
 {
     return argsSize_;
